# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/{path}":{"options":{"summary":"Cors Preflight","operationId":"cors_preflight__path__options","parameters":[{"name":"path","in":"path","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/weeks/current":{"get":{"summary":"Weeks Current","operationId":"weeks_current_weeks_current_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WeekOut"}}}}}}},"/weeks/{week_id}/songs":{"get":{"summary":"Weeks Songs","operationId":"weeks_songs_weeks__week_id__songs_get","parameters":[{"name":"week_id","in":"path","required":true,"schema":{"type":"integer","title":"Week Id"}},{"name":"filter","in":"query","required":false,"schema":{"enum":["all","new"],"type":"string","default":"all","title":"Filter"}},{"name":"search","in":"query","required":false,"schema":{"type":"string","default":"","title":"Search"}},{"name":"x-telegram-init-data","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Telegram-Init-Data"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SongOut"},"title":"Response Weeks Songs Weeks  Week Id  Songs Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/weeks/{week_id}/results":{"get":{"summary":"Weeks Results","operationId":"weeks_results_weeks__week_id__results_get","parameters":[{"name":"week_id","in":"path","required":true,"schema":{"type":"integer","title":"Week Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/weeks/{week_id}/vote":{"post":{"summary":"Weeks Vote","operationId":"weeks_vote_weeks__week_id__vote_post","parameters":[{"name":"week_id","in":"path","required":true,"schema":{"type":"integer","title":"Week Id"}},{"name":"x-telegram-init-data","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Telegram-Init-Data"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoteIn"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoteOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/weeks/current/songs/bulk":{"post":{"summary":"Admin Add Songs","operationId":"admin_add_songs_admin_weeks_current_songs_bulk_post","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{},"title":"Songs"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/weeks/current/songs/enrich":{"post":{"summary":"Admin Enrich Current Week","operationId":"admin_enrich_current_week_admin_weeks_current_songs_enrich_post","parameters":[{"name":"x-admin-token","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"X-Admin-Token"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"boolean","default":false,"title":"Force"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/__version":{"get":{"summary":"  Version","operationId":"__version___version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SongOut":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"artist":{"type":"string","title":"Artist"},"is_new":{"type":"boolean","title":"Is New","default":false},"cover":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cover"},"preview_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Preview Url"},"source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source","default":"manual"}},"type":"object","required":["id","title","artist"],"title":"SongOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VoteIn":{"properties":{"song_ids":{"items":{"type":"integer"},"type":"array","title":"Song Ids","default":[]}},"type":"object","title":"VoteIn"},"VoteOut":{"properties":{"ok":{"type":"boolean","title":"Ok"},"week_id":{"type":"integer","title":"Week Id"},"user_id":{"type":"string","title":"User Id"},"voted_song_ids":{"items":{"type":"integer"},"type":"array","title":"Voted Song Ids"}},"type":"object","required":["ok","week_id","user_id","voted_song_ids"],"title":"VoteOut"},"WeekOut":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"status":{"type":"string","enum":["open","closed"],"title":"Status"}},"type":"object","required":["id","title","status"],"title":"WeekOut"}}}}